+++ htfs/virtual.go
@@ -47,7 +47,7 @@ func (it *virtual) Remove([]string) error {
 	return fmt.Errorf("Not supported yet on virtual holotree.")
 }
 
-func (it *virtual) Export([]string, string) error {
+func (it *virtual) Export([]string, []string, string) error {
 	return fmt.Errorf("Not supported yet on virtual holotree.")
 }